Split up compiler module
